--- a/swiftsimio/visualisation/slice.py
+++ b/swiftsimio/visualisation/slice.py
@@ -418,13 +418,15 @@ def slice_gas_pixel_grid(
     hfinal = array(hsml / max_range)
     rescaled_box = array([box_x / max_range, box_y / max_range])
 
-    xall = array([])
-    yall = array([])
-    zall = array([])
-    mall = array([])
-    hall = array([])
+    xall = xfinal.copy()
+    yall = yfinal.copy()
+    zall = zfinal.copy()
+    mall = mfinal.copy()
+    hall = hfinal.copy()
     for xshift in [-1, 0, 1]:
         for yshift in [-1, 0, 1]:
+            if xshift == 0 and yshift == 0:
+                continue
             thisx = xfinal + xshift * rescaled_box[0]
             thisy = yfinal + yshift * rescaled_box[1]
             inside = (