--- a/swiftsimio/visualisation/volume_render.py
+++ b/swiftsimio/visualisation/volume_render.py
@@ -395,14 +395,16 @@ def render_gas_voxel_grid(
     hfinal = array(hsml / x_range)
     rescaled_box = array([box_x / x_range, box_y / x_range, box_z / x_range])
 
-    xall = array([])
-    yall = array([])
-    zall = array([])
-    mall = array([])
-    hall = array([])
+    xall = xfinal.copy()
+    yall = yfinal.copy()
+    zall = zfinal.copy()
+    mall = mfinal.copy()
+    hall = hfinal.copy()
     for xshift in [-1, 0, 1]:
         for yshift in [-1, 0, 1]:
             for zshift in [-1, 0, 1]:
+                if xshift == 0 and yshift == 0 and zshift == 0:
+                    continue
                 thisx = xfinal + xshift * rescaled_box[0]
                 thisy = yfinal + yshift * rescaled_box[1]
                 thisz = zfinal + zshift * rescaled_box[2]