--- a/swiftsimio/visualisation/volume_render.py
+++ b/swiftsimio/visualisation/volume_render.py
@@ -254,6 +254,7 @@ def scatter_parallel(
 
     return output
 
+
 def render_gas_voxel_grid(
     data: SWIFTDataset,
     resolution: int,