The visualisation functions only consider one periodic copy of each particle. As a result, there are some weird edge effects when visualising the full box. This is what you get when you visualise the full `cosmological_volume.hdf5` test file ("zoom" is the region inside the white square that excludes the edge):

Projection:
![project](https://user-images.githubusercontent.com/7336967/181795238-ab375f16-4e2d-4e8c-9149-39e944ee6e9d.png)

Slice:
![slice](https://user-images.githubusercontent.com/7336967/181795260-69821f96-47d8-447a-bd17-48a2a89ec0a4.png)

Volume render (slice through the centre of the voxel cube):
![render](https://user-images.githubusercontent.com/7336967/181795289-849eaffd-1c32-4877-a01b-821829cad0e1.png)

This PR adds periodic copies of all particles that could potentially contribute before calling the visualisation function. This restores the correct edge:

Projection:
![project](https://user-images.githubusercontent.com/7336967/181795615-72246f82-d3e2-450b-a9cd-f8fa40a4756e.png)

Slice:
![slice](https://user-images.githubusercontent.com/7336967/181795651-5bf61fbf-434b-43c4-9abe-417a9996de6e.png)

Volume render:
![render](https://user-images.githubusercontent.com/7336967/181795683-d3de8dcb-83b5-4e9c-8630-3888adf72860.png)