@JBorrow 2022-07-29
This change looks like it may introduce significant performance impacts. Is there some way of checking this?