@JBorrow 2022-07-29
I would also suggest it may be quicker to just run the vis code twice (one for periodic particles, one for non-periodic) than append to the arrays.