@bwvdnbro 2022-08-01
Of course there will be a performance impact. We are including more particles.

Some quick timings based on the scripts I used for the images above (times are the mean and standard deviation for 10 successive calls to the corresponding visualisation function):
```
master:
project:
Took 10732.50 +- 230.36167937716385 ms
slice:
Took 357.06 +- 198.60511500139805 ms
render:
Took 832.70 +- 227.8507330442371 ms

new branch:
project:
Took 12704.47 +- 222.1358350366033 ms
slice:
Took 392.93 +- 181.07175431095246 ms
render:
Took 1062.95 +- 210.43395451782126 ms
```
So roughly ~20% more.

I don't think appending to the arrays is particularly slow, because this only happens 9 or 27 times and scales linearly with the number of particles. The runtime is generally dominated by the scatter function. I could eliminate the iteration of the loops that deals with the normal case (`x/y/zshift==0`), but that would make the code uglier for probably little gain.
Calling the scatter function twice has the disadvantage that it introduces an extra synchronisation point if you are using a parallel scatter function. So you increase the potential for dead time because of load imbalances.
The only good reason I can see to call the scatter function twice (or even more) is to reduce the memory footprint. That will be larger in the current implementation and is generally unpredictable (worst case, it could be 27 times larger for the volume rendering). But I doubt this will be a serious problem for any real application.