@JBorrow 2022-08-02
Does it introduce another sync point? The addition to the main array, as far as I know, is actually handled asynchronously.