@JBorrow 2022-08-02
You can also do it in parallel, add scatter_inside + scatter_outside then add this to the main array?