@JBorrow 2022-08-02
Hmm, looking at the code I see what you mean about sync point. But this should be well load-balanced, anyway.