@bwvdnbro 2022-08-02
> Does it introduce another sync point? The addition to the main array, as far as I know, is actually handled asynchronously.

I might be misunderstanding what you meant then. I thought you were suggesting to have multiple calls to
```
    if parallel:
        image = backends_parallel[backend](**common_arguments)
    else:
        image = backends[backend](**common_arguments)
```
This is not done in any parallel bit of the code, so at the end of each call you would get a new synchronisation point.

Anyway, I have tested this and is definitely slower than calling the backend once with the appended arrays.

What does seem to make a difference is initialising the arrays with the particles and skipping the `x/y/zshift==0` case in the loop. At least for the projection. For the slice and volume rendering this is actually slower and I don't understand why.