@JBorrow 2022-08-03
I would like to see this trailed on something a bit bigger before merging.

In particular, I'm concerned about memory consumption for the volume rendering, as this is already extremely memory-intensive (though that mainly comes from allocating the voxel grid).