@bwvdnbro 2022-08-04
I have run the projection code on a thin slice through one of the 1 Gpc FLAMINGO snapshots and the results are not great. The memory footprint and run time increase quite dramatically. I'll try some more things.