@JBorrow 2022-08-04
Is it worth changing the numba functions to check if the particle overlaps the edge, then seeing if it would partially project onto the other side? This way you would only have one copy of data in memory, and branch prediction would likely take care of any negative performance impacts?