@bwvdnbro 2022-09-08
I was hoping to avoid touching the numba functions, so that this is implemented in one place instead of in every single backend. But I guess that will indeed be a better solution given the impact on the performance.

I would like to keep this branch alive for reference until I get a chance to properly implement this in the backends. Having a slow but working version is useful for testing.